requirements.txt
numpy==1.16.4
Pillow==5.4.1
scikit-image==0.14.2
torch==0.4.1
torchvision==0.2.1
tqdm==4.28.1
